Refactor Vmware Cloud manager by moving auth functions to mixin
Since NetworkManager will use same auth functions as CloudManager,
we move them into mixin that both will be using. Functionality
should remain the same.
